by ricardo vitorino (rjfv): bugfix about private thread
"When a private thread was opened and the user was not logged in or wasn't
the administrator, it used to return an HttpResponseNotFound. Not only the
HttpResponseNotFound should be "raised" instead of "returned", but also,
it should be an error like Permission Denied or Unauthorized."

i merged rjfv's commit to updated develop branch.